Creacion de widget central con informacion del cliente y pantallas segun acciones a ejecutar del cliente
